Ensure we can access keys on request.headers
Previously, we converted requests' CaseInsensitiveDict [0] to a dict
which preserved the data structure of CaseInsensitiveDict. This
created a top-level _store key, with an interior OrderedDict that
had the headers we want to access. Instead, we should use the
dict() attribute which gives us just the headers and not the outer
_store and OrderedDict.

[0] http://docs.python-requests.org/en/master/_modules/requests/structures/