Database: Store UUIDs of observed journalists
In a conversation flow, there may be one source and multiple
journalists. In order to display the usernames of journalists,
we must save locally the journalist UUID and last observed username.

Note that usernames can change, so each time we synchronize data
from the server, we should update the username if it has changed.