data.py: Gracefully handle missing files on disk
So, I can modify gui/main.py, but I'd need to refresh the source
object which requires access to the current session.

I think the cleanest thing to do here is to make sure that if the
file is missing on disk, then the application doesn't crash, and
instead the message should indicate that it's been deleted.