replies: temporary fix to prevent exception in happy path of app
Previously, an exception would occur due to self.api being None
when a user is not logged in. Since we're going to address this
soon by, e.g. greying out the reply box when we're not logged in,
let's disallow sending the reply if the journalist
is not logged in. We'll also want to emit the failed signal to
let the developer know what has happened.