demonstrate our segfault is an upstream bug: QTBUG-69204
Appearing in the core dump for the seg fault was the following:

Thread 0 Crashed:: Dispatch queue: com.apple.main-thread
[...entirely obtuse part snipped out...]
2   QtWidgets                     	0x000000010a6bb4ef QWidgetPrivate::setStyle_helper(QStyle*, bool) + 639
3   QtWidgets                     	0x000000010a6bb239 QWidgetPrivate::inheritStyle() + 457
4   QtWidgets                     	0x000000010a6b5e39 QWidget::setParent(QWidget*, QFlags<Qt::WindowType>) + 1705
5   QtWidgets                     	0x000000010a6a8d43 QLayout::addChildWidget(QWidget*) + 211
6   QtWidgets                     	0x000000010a6a02ce QBoxLayout::insertWidget(int, QWidget*, int, QFlags<Qt::AlignmentFlag>) + 62
7   QtWidgets.so                  	0x000000010a3475a5 meth_QBoxLayout_addWidget(_object*, _object*, _object*) + 197

Searching the QTbug tracker for issues involving setStyle_helper, we find
an ongoing upstream bug in the handling of QStyleSheetStyle's updateObjects()
causing segfaults [0].

If this is indeed the problem we are hitting, then we should be able
to disable all our CSS styling and everything should work, which
is what is done in this commit, demonstrating that we're hitting
this upstream bug.

This commit also adds lots of extra logging I was using for figuring what
was going on here.

[0] https://bugreports.qt.io/browse/QTBUG-69204