0.1.3: Update PyYAML to 5.1
Also simplify package versioning with a VERSION file containing the
version string, e.g. 0.1.3, which will be read in setup.py and
securedrop_proxy/version.py. Now we just have to remember to update
the package version in one place.

Add a test making sure that the proxy reports the same version used in
setup.py.