Resolve review comments
Revert the reversion of API.token to the dict received from the
server. Instead, keep API.token as the token itself, restore
token_expiration as a datetime, and add token_journalist_uuid. The
client will have to be updated to get the UUID from that attribute
instead of extracting it from API.token.

Keep the change to return True from API.authenticate, but indicate
problems with exceptions.