inject session_maker's into objects instead of passing Session's and SQLAlchemy objects
Currently we have managed to avoid threading issues with SQLA. Passing a
Session or Model (object corresponding to a DB row) across thread
boundaries is dangerous and will raise exception. We have managed to
avoid problems with this so far because most of our objects live in the
main GUI thread. Additionally, our signal/slot connections all live in
the main thread, and when they don't, we might actually still be
triggering actions in the main thread because we're not using a
Qt.QueuedConnection when we connect them.

This commit attempt to give objects their own session_maker so they can
create sessions on demand and fetch thread-local objects from the DB.