Extract and display original document filenames
Files submitted to the SecureDrop server are gzipped with their
original name, then encrypted with GPG. When downloaded, we can
extract the original filename from the gzip header, so they can be
opened with an appropriate application.

This change adds the db.File.original_filename column, and populates
it when downloading and decrypting files. The file is still stored as
it has been, under the db.File.filename value stripped of extensions,
but when opened, that file is first hard linked to the original
filename.

Since we can't wait for the file to be opened (starting the disposable
VM is pretty slow), we can't immediately remove that link, so
storage.delete_single_submission_or_reply_on_disk now cleans it up if
the object being deleted has original_filename populated.