queue: jobs must be sortable, i.e. must implement __lt__
the next job in a priority queue is set by sorted(list(entries))[0]

this means that the job objects themselves must be sortable, i.e.
they must all implement __lt__.

in this commit I'm just implementing __lt__ for two job types where
we don't care about the order as long as all jobs of that type have
the same priority.

see related discussion in https://bugs.python.org/issue31145 and for
a solution using dataclasses (not in python 3.5)