Updates run.sh for Qubes compatibility
Infers whether the developer's host OS is Qubes, based on the presence
of env vars matching pattern `^QUBES_`. If Qubes, we can permit the use
of DispVMs for opening submissions.

The logic assumes that the Qubes AppVM running the client (e.g.
`sd-dev`) has the tag `sd-client` applied. We cannot inspect this from
within the AppVM, so we'll lean on documentation to instruct developers
about this.