bugfix: populate source list immediately upon login (#577)
the controller's update_sources() method is used to display in the
source list whatever is contained in local storage. If we wait until
after sync_api is called before calling it, there will be a delay
of potentially several seconds until the round-trip to the SecureDrop
server completes.

in this commit we call the update_sources() method immediately
upon login.