Improve error handling, tests
Try to ensure the proxy always returns a valid JSON response. Add
specific handling of upstream error responses to proxy.py, and add
tests and fixtures for that.

Replace the Proxy._on_done static method with an instance method.

In main.py, bail out upon receiving invalid input JSON, instead of
printing an error response and then continuing to try to use it,
resulting in printing another error response.

Have main.py use the callbacks on the proxy as given, instead of
always overwriting them with the defaults from callbacks.py.

Rework entrypoint.py so that it should always print a JSON response,
and add tests for it.