Improve download file handling
Eliminate File.original_file column and hard links to downloaded
files. Now as files are downloaded and decrypted, File.filename will
be updated.

Instead of downloading directly into the root of the data directory,
create a directory therein for each downloadable object, named with
its UUID, and process downloaded files in that directory. This should
eliminate collisions if multiple sources upload documents with the
same names.