Use journalist designations, not UUIDs, in data directory
Base directory and file names in the data directory on the journalist
designation of their sources, plus file counters and extensions,
instead of UUIDs.

Remove handling of changes to source journalist designations on the
server; local files will no longer be renamed. The ability to change
journalist designations will soon be removed from SecureDrop core, so
this will not be needed.