WIP -- demonstrating problems with transactions
This modifies storage.__update_submissions to only try to create a
submission if its source exists. This would solve the scenario posited
in #797, but it shouldn't. If transactions are working properly, the
deletion happening on the GUI thread should not be visible within the
sync thread's call to update_local_storage; the source should still
exist when the new message is created. So if transactions were
functional, this change would not be necessary.

To prove this, uncomment the "do_begin" event listener in
db.make_session_maker, then run:

    make test TESTS="tests/test_storage.py::test_sync_delete_race"

It should fail when asserting that the source has been deleted.