Clearer message in "no keypair" case, fix styling
Both ReplyBoxWidget and ReplyTextEdit must be disabled when no
keypair is available for a source. I have changed the method
names to `set_logged_in`/`set_logged_out`, consistent with
the use in ReplyTextEdit, because this more accurately describes
what they do (the widget can be in online mode, but disabled).