Work around problems with Qt stylesheet handling in widget hierarchy
Simply setting the stylesheet on ReplyWidget has stopped working;
styles are not being applied to children. My working theory is that it
might be due to changes in the Debian package qtbase-opensource-src at
5.11.3+dfsg1-1+deb10u2, which incorporated upstream changes that
changed how stylesheets are applied to children. That theory is a bit
shaky, though, because with the same system packages installed,
installing the same version of PyQt5 in a virtualenv results in the
old approach working fine. That's unfortunately not an option for us
at this time, as it would be a huge addition to the production
requirements, and potentially painful to keep in sync with the system
Qt packages.

At any rate, the workaround in this change, to explicitly apply the
CSS to the children, seems to solve the problem of reply widgets not
being updated correctly for their status.