app: when calling SourceList.clear(), remove the references in the dict
We previously were keeping references to pyqt Python wrapper objects
that had their corresponding C++ objects deleted.

This means that if we call any qt method on the Python object, we'd
get a RuntimeError indicating that the underlying
object is gone. We should ensure that we remove references to the
Python wrapper object, so that we do not cause the aforementioned
RuntimeError (which crashes the application)
and/or prevent these wrapper objects from getting garbage collected
as usual by cpython.