Fix StarToggleButton hover leave
The icon was being reset to always-on even if the source weren't
starred. This adds a check for that, and also corrects the misleading
naming of the attribute determining whether the star API calls should
be made when the widget state changes. (If it only takes a weekend for
your naming to start tripping you up, it was clearly not thought
through.)