Clean up cleanup in run.sh
The cleanup function was busier than it needed to be, and because of
the exec of the client, was never actually being invoked. This change:

- reworks cleanup to use "gpgconf --kill" to stop the agent
- invokes it explicitly before the client is run
- removes the exec so that "trap cleanup EXIT" works and cleanup
  happens after the client exits

Making sure gpg-agent is stopped should prevent any problems it might
have if GNUPGHOME is pulled from under its feet between client runs.

This should make run.sh more robust when developers are mucking about
with SDC_HOME.

We might consider adding similar agent cleanup in production, but
since SDC_HOME should be consistent and in normal usage should never
be removed, the problems we've seen with the GPG agent should not
arise in production use.