logging: include uuid in __repr__, do not include File.filename
since we use the __repr__ in log lines, we should exclude the
File.filename from File.__repr__ such that we do not log the _real_
filename of a source document. For messages and replies it's fine
to log the filename as it'll just be the journalist_filename.

this commit also modifies some log levels: info and above will be
logged in prod, debug for dev.