Speed up update_local_storage in sync
Stop importing source GPG keys during sync. We have a place on Source
to store fingerprint and key, which is all we need for the import, and
which is what we check when deciding whether to enable the reply box,
so let's just update the database here and defer import to the keyring
until someone actually wants to reply. It takes milliseconds that
won't be noticed then, but do add up during sync.

Avoid dirtying synced objects on update unless they've actually
changed.

Use maps instead of sets to hold local objects in the update_
functions, so it's faster to check if we already have a record of
incoming objects.

Reduce the number of database queries during sync: cache sources or
journalists instead of looking them up for each incoming object
associated with them.