Improve error handling of downloads, speech bubbles
Adds error state to SpeechBubble, so messages and replies that fail
to download will change style and display the error message within.
They should automatically recover when errors clear.

Add db.DownloadError with relationships to File, Message and
Reply. Use that to determine styling and content of messages and
replies in the conversation view, instead of signals, because if the
conversation view has not been built, there's nothing listening for
the signals, and when the source is clicked, the styling and text
won't reflect the current downloadable object state.

Note that while File can now have a download error associated,
file widgets are not yet using it.

Add last_updated column to File, Message and Reply, and use that in
conjunction with download_error to delay retry of downloads that fail
due to decryption problems, to avoid sd-gpg access notification spam.