Modify scripts to run only in sd-app
In order to support template consolidation, we should prevent the client from opening on VMs that are not named `sd-app` as well as customize the qubes-gpg-domain to sd-gpg. This should not affect any dev environments, since they rely on the `run.sh` script, and this only applies when invoking `securedrop-client` and through the .desktop file, which is only provisioned through the .deb file.

While the securedrop-client AppArmor profile is enforced by default, changes to the client AppArmor profiles will not be applied. We need to run `aa-enforce` on the executable, but we should only do it in `sd-app` to ensure AppArmor kernel support is present.