Moves config to private volume
The Qubes RPC file hardcodes the filepath to the YAML config file,
which contains site-specific information such as the Onion URL for the
Journalist Interface. As part of template consolidation [0], we're
moving the config file out of the system/root partition and into the
private (i.e. /home/) volume, so that the `sd-proxy` AppVM has the
config information it needs while sharing a TemplateVM with other
components.

[0] https://github.com/freedomofpress/securedrop-workstation#471