Fix type annotations, use explicit NoReturn
Specifying when functions do not return allows Mypy to warn
when unreachable code is introduced.

Note: that feature is enabled by the option --warn-unreachable
which is not currently used. I also noticed that enabling type
ceking on this file uncovers errors in the GUI code, so I'll
address those before modifying the Mypy options.

See https://www.pyhton.org/dev/peps/pep-0484#the-noreturn-type