Refactor LoginOfflineButton as QPushButton
The Qt buttons do integrate with the operating system, as a result
they are easier to take advantage of for accessibility purposes.

This is a minimal modification: the button behavior has not changed,
even though it could be improved in a few aspects (e.g. implementing
some of the conventional button states would allow to provide timely
feedback to the person interacting with the button).

Decisions:

- I called the base class SDPushButton by analogy with QPushButton.
  The intention is for that class to provide a generic button, adequate
  to the SecureDrop Client context.
- I'd argue that LoginOfflineLink wasn't a link - understood as
  hyperlink. Links come with expectations of behavior (e.g. history)
  that are not met in this case. On the other hand, the behavior of
  LoginOffineLink was close to that of a dialog "dismiss" button.
- I didn't write any test for SDPushButton because it doesn't implement
  any behavior that was tested in LoginOfflineLink. In other words,
  the button is not less tested as it is.
- I made a stylesheet specifically for the button so that the styles
  are scoped to the button. This should make maintenance easier.

See https://material.io/components/buttons
and https://speakerdeck.com/didoo/let-there-be-peace-on-css?slide=62