Fix remove inaccurate message when source is selected
The conversation area can be overlayed with different
messages depending on the circumstances. If no sources
are selected, a given overlay invites to select one.
If all messages from the conversation have been deleted,
a distinct overlay indicates so.

The former overlay is visible when created, then hidden
when not necessary, typically when a source is updated.

That update isn't being triggered after logging in
from offline mode, though, which resulted in both overlays
being superposed. Gladly, they aren't completely opaque
which made the bug apparent.

There is no need to fully update the conversation,
as long as we hide the "empty converation" overlay
if a source is already selected.

Co-Authored-By: Allie Crevier <allie@freedom.press>