Enable keyboard-only interactions for increased accessibility
There is currently only one button in the dialog, so it is already
the default button. However, I think it's worth declaring it
explicitly as such.

Enabling the default 'Escape' key behavior provides a way to
close the dialog and return the focus to the main window.
If the main window is closed, closing the dialog quits the application.