Fix mistaken default button in login dialog
The buttons can be defined as "default button", ONLY when they
are children of a QDialog. That was the missing part.

Note that both buttons are "auto default", that is why it is
necessary to specify which one should take precendence by
calling its setDefault method. Otherwise, the rendering order
of the buttons would determine which one is activated when
the journalist presses the Enter key.

See https://doc.qt.io/qt-5/qdialog.html#default-button
and https://doc.qt.io/qt-5/qpushbutton.html#default-prop