Refactor remove widgets from package init file
Moving these widgets to their own module allows to prevent
circular dependencies when moving code from the 'widgets' module
to their own file.

Please bear with the 'misc' module name, it is intentionally vague
and I intent to find a proper place for all its contents as part
of the current refactoring.

Noteworthy:

- Creating widgets in __init__.py meant iporting some of their
  dependencies, e.g. QHBoxLayout.
- Those dependencies would find themselves exported by the package :S
- And as a result, the tests were mocking QHBoxLayout.addWidget
  as if it was defined in securedrop-client.gui !
  Strictly speaking, it was indeed re-defined there, but I'd say
  that this in itself a good reason not to write widgets in __init__.py.