Refactor encaspulate SDMenuBar key settings
There is no need for the main window to know in advance
about which specific key event toggles the menu bar.

Note that another option would be to forward all key events
to the menu bar, but I think that minimizing the number of
events that the menu bar has to process is worth the trouble
in this case.