Fix flashing menu bar on Alt+TAB (generally: key combinations)
As reported by @eloquence, when performing an Alt+TAB combination
to switch windows, the menu bar would become briefly visible (before
it lost focus and got hidden again). For reference, that doesn't
happen in Firefox.

My understanding is that when performing a combination, the Alt key
press event happens first. If the menu bar is listening to it, it
becomes visible. Then, when the TAB key event happens, the focus switches
to the other window. Finally when the Alt key release event happens,
the menu bar doesn't receive it because it is not in focus anymore.

Toggling the visibility of the menu bar on key *release* events
therefore solves the issue, and I confirmed that it behaves like
Firefox does now when Atl+TAB is performed.

Additionally, the Escape key behavior changes slightly, also to become
the same as in Firefox: when a menu was open, pressing Escape would
close the menu, and pressing it again would toggle the menu bar.
The menu and the menu bar are now closed and toggled at once.
(I don't think it makes a significant difference, but I hadn't
noticed that Firefox behaved differently, so the convergence is good.)