Refactor simplify buttons creation
As @creviera pointed out in #1339, the buttons' parent is set when
those are added to a layout (via addWidget).

The setDefault method must be called _after_ the button's parent
chain was established to make sure that the QDialog in that chain
actually makes the method effective. (That hasn't changed.)

Delaying setDefault and relying on the automatic parent assignment
is simpler, but I now realize that it is also less confusing.
The parent of the buttons will eventually become the layout they
are added to, and manually setting the QDialog (self) as their parent,
was causing a change in parenting a few lines below.

While that doesn't interfere with the behavior we want, it's certainly
unnecessarily confusing! Thanks @creviera for pointing that out!

I removed the parent parameter from the constructors of the two buttons,
because their roles are well-known and we shouldn't need it.

I kept it in the constructor of SDPushButton, because it is a more
generic widget and the case is less clear-cut. We might end up removing
it too, but I'd keep it for now.