CI: Switch to debian:buster image
Circle CI is updating the images they maintain in Docker Hub [0].
Ostensibly we're supported to switch from `circleci/python:3.7-buster`
to `cimg/python`, but the latter image is based on Ubuntu Focal,
whereas we have a hard requirement for Debian Buster (until we migrate
to Debian Bullseye). So instead let's use `debian:buster`, which
is not a CircleCI image, but fits our use case.

Some changes are required: make sure sudo is present, update paths
not to assume a normal user, install xvfb. Otherwise, fairly
straightforward.

[0] https://circleci.com/docs/2.0/next-gen-migration-guide