venv-mac: installs unpinned dependencies, without hashes
This is a compromise between (a) having to maintain a full
"dev-mac-requirements.txt" and (b) not providing *any* support
for development on macOS.  In effect, by installing dependencies
without "--require-hashes", we hereby offer macOS support on a
best-effort basis, to be used by contributors at their own risk.

See discussion in #1396.