Ensure the state is initialized...
...from the persitance layer when the application is started.

Eventually, the GUI could depend exclusively on state.State,
and a method could be added to persist it to the database.

That would remove the necessity to setup a database in order to
use the GUI. When present, such a database would allow state to
persist across restarts, which is exactly its role.

In particular, the database would not be responsible for transporting
information between different parts of the GUI. That repsonsibility
would be that of state.State, which can have whatever internal structure
is best suited to the task, without any constraint related to database
data layout.

This can be done progressively, by expanding the Database collection
of methods as convenient.