Fix menu sub-sections vertical spacing for readability
Ensure that the menu sub-section titles are closer to the sub-section
items than they are the last item is to the next sub-section title.

Ensure that the action items are clickable from one edge of the menu
to the other (like standard menu items).

Increase the menu items size (area).

See https://lawsofux.com/law-of-proximity
and https://lawsofux.com/fittss-law