fix: ApiJobQueue.stop() waits for processing to finish and clear queues
Here we understand stopping an ApiJobQueue to be a synchronous operation
in which we:

1. inject a PauseQueueJob to suspend the processing of subsequent jobs;

2. stop the ApiJobQueue's two processing threads; and

3. wait for those threads' "finished" signals to trigger each
   RunnableQueue's clear() slot.

In (1), why can't we interrupt the processing of the current job?
Because RunnableQueue.process() blocks the native QObject event loop[1],
we can't trigger the clear() slot directly.  Because process() blocks on
waiting for a non-empty underlying PriorityQueue, we can't interrupt it
at the thread level with RunnableQueue.thread().requestInterruption().

[1]: https://forum.qt.io/topic/92932/how-to-process-events-during-infinite-loop-in-worker-thread/2