fix(mark_all_pending_drafts_as_failed): filter for replies pending under previous PIDs
If a DraftReply was marked "pending" by a previous Client session (i.e.,
with a different PID), by now it must have either failed or been sent
successfully.  To avoid leaving it an an inconsistent state, we presume
it failed; we'll update it anyway if the Journalist API reports it
received on the next sync.