Refactor ensure signals are used as recommended
The controller used to imperatively emit signals
on the Export instances. This breaks the encapsulation
provided by the Export class and defeats one of
the main benefits of using signal: de-coupling components.

From Qt's documentation:

    Signals are public access functions and can be emitted
    from anywhere, but we recommend to only emit them
    from the class that defines the signal and its subclasses.

The controller now emits a signal indicating that
a preflight check is desirable (for example) and the Export
instance connects its own slots to that signal.

See https://doc.qt.io/qt-5/signalsandslots.html