Rephrase pull request template guidance to remove checklist
The check lists are picked up by GitHub as tasks to be completed,
which makes "radio"-style checklists misleading: some items
that GitHub identifies as tasks will never be completed!

While this is a matter of GitHub's interpretation, I think it is
useful to take advantage of the summary for evaluating how much
work is remaining on a PR.

Besides, the items were phrased as if tasks performed by
the pull request authors didn't need to be performed again
by the reviewers, which is not how we perform reviews.