Ensure application lifetime is clear in integration tests
The main application thread lifespan is easier to follow when
the application instance is created and destroyed for each
integration test. The current use of the exisiting fixtures
allows that (only one of them is used in each test), so that's
what I did.

This is part of an effort to ensure that threads are closed
properly before their Qt/Python references are destroyed.
Even though failure todo so doesn't always result in a failed
test suite, it will become easier to do well if we do it consitently.