Mark one more functional test as flaky
This isn't the most elegant solution, but it seems to me like
it's a reasonable and pragmatic one in today's context.

There is little penalty for marking functional tests as flaky,
and it is less time-consuming (and probably more platform-agnostic)
than finding out the correct timeout for any given operation.

	[...]
	Special cases aren't special enough to break the rules.
	Although practicality beats purity.
	[...]
	-- The Zen of Python