style: reorder table and column operations to minimize SQL diff
The SQL statements generated by "sqlite3 svs.sqlite .dump" replicate the
database's tables and columns in the order in which they were originally
added.  For the current database schema, that's a hybrid of the order in
which they're defined in securedrop_client.db and subsequent additions
in migrations; in the new database schema, that's strictly their
ordering in securedrop_client.db.  The latter artificially inflates the
diff of comparing "sqlite3 svs.sqlite .dump" for the old and new
schemas.

For ease of review, here we attempt to replicate the current SQL
statements as closely as possible, for as small a diff as possible.