Refactor inject export.Service into main window
The export service is currently a dependency of the export dialog.

Treating it as such allows to simplify testing by making possible
to control the characteristics of the export service that's used
in testing context. Namely, this commits introduces a new fixture
that allows to test the export dialog without relying on Qubes OS,
and without relying on mocking of implementation details.