Fix dependency injection semantics, make export_service required
The export.Device does depend on an export.Service instance.

Injecting an instance which runs in a separate thread is desirable
to preserve GUI responsiveness during the (potentially long) export
operations. That requires carving an injection path through multiple
other widgets.

I choose to declare the export_service as Optional in those cases
to mean that their specific behavior doesn't depend on it.

Thanks @rocodes for pointing out the Optional overreach!