chore: export Qubes service flag "SDEXTENDEDTIMEOUT_N" as SDEXTENDEDTIMEOUT=N
Since a Qubes service just sets a boolean flag, we use qubesdb-list (1)
as a glob, for any key beginning with the prefix "SDEXTENDEDTIMEOUT_";
and (2) to return the "value", aka the key without that prefix.

This approach is too naïve to work for arbitrary keys (which risk
colliding in the glob) or arbitrarily-typed values (without encoding).
But it's good enough for this experimental flag.  More-sophisticated
approaches, such as scanning the contents of "/var/run/qubes-service",
would require AppArmor grants, which we already have for qubesdb-cmd.