Fix missing dependency: py
In the previous iteration, py was ostensitbly beinginstalled as a
dependency of pytest and pytest-forked.

The package pytest-forked is no longer installed, because the new
version of pytest doesn't depend on it anymore.

It doesn't seem like pytest declares py as a dependency anymore either,
because its vendors the parts it requires. (Change log linked below.)

However, it seem that we do otherwise require py, so I added it.

Note: the Bullseye and Bookworm requirement files were generated using
Python 3.9 and 3.10 respectively.

See https://docs.pytest.org/en/stable/changelog.html#pytest-7-2-0-2022-10-23