Allow development environment to work on arm based macOS machines
Fixes #1591

This also changes the install instructions on macOS in general, unifying
the setup for Intel and arm64 machines. This also drops instructions to
use pyenv in favour of installing Python 3.9 via Homebrew. This is
necessary to allow x86_64 mode Python 3.9 virtual environments to work
as we need them to.