Increase amount of allowed retries of flaky tests
This is in no way making the test suite better. However, given
that we're running CI pipelines again regularly, allowing those
retries to happen automatically will save time.

There is no doubt that we want to fix the underlying issue(s)
eventually, but have decided not to do it immediately.