Refactor make export service a module instance
This will ensure that one one instance of the export service
queries the sd-devices CLI.

That was already the case, but nothing was done to suggest it.

Additionally, this will allow to simplify the export service
injection by making possible to retrieve it like a logger.

Note: I haven't stopped injecting the export_service into
the main.Window because I couldn't figure out how to prevent
what seems to be a fixture scope error that affects only
half of the integration tests. That seems like something that
can be resolved, so I don't consider it to be a sign of
export.Service being a bad idea.