Refactor deprecate export.Export test file
I'll use an expand and contract pattern to refactor
the export.Service tests, and keeping this file distinct
while it is used as a reference will be more convenient
and minimize the change sets.

Important: this file is still fully operational and will be
kept as such until all deprecated methods have been removed
from export.Export (soon to be renamed export.Service).