fix(ReplyWidget): handle update_signal as successful reply
Fixes a partition[1][2] between SpeechBubble._update_text(), called when
a reply has been downloaded from the Journalist API, and
ReplyWidget._on_reply_success(), called when a reply has been
successfully sent via the SDK.  In the case where the Client starts with
a reply still pending from a previous session and the reply is
subsequently downloaded, ReplyWidget.status will not be updated from
Reply.send_status unless it too listens for the SpeechBubble's
update_signal, and future calls to ReplyWidget._update_styles() will be
based on a stale ReplyWidget.status.

While conflating these signals is not ideal, in the current data model
we might as well, since a reply downloaded from the server is by
definition one that has been successfully sent to the server.  See
<https://github.com/freedomofpress/securedrop-client/issues/1493#issuecomment-1130898792>.

[1]: https://github.com/freedomofpress/securedrop-client/pull/1486#pullrequestreview-1190947065
[2]: https://github.com/freedomofpress/securedrop-client/pull/1486#issuecomment-1334572068