Add semgrep rule for unsafe assertions
Assertions are not checked in production, and are risky to use
for any reason that relies on them at runtime.

Type checking is not a runtime concern, and using assertions
to narrow types during type checks is not a problem.

See https://mypy.readthedocs.io/en/stable/type_narrowing.html
and https://snyk.io/blog/the-dangers-of-assert-in-python/