Refactor remove deprecated methods, rename for consitency
Rename disk functions in export.CLI and export.Service
Rename printer functions in export.CLI and export.Service

The CLI now encapsulates all "preflight" references, so that
neither the CLI or the Service expose them publicly.
(References remain on the export.Device and dialogs, but those
can now be removed for consitency with the export.Service.)

Refactor remove unnecessary methods that were already deprecated

Note: some of the deprecated tests were testing private methods
that are currectly CLI behavior. This commit moves that testing to
the appropriate file.