Refactor remove unnecessary export.Service injection
The export.getService() function allows to obtain the export service
instance wherever necessary without needing to inject it through all
the parent widgets.

This is a double-edged sword, as it makes the injection more implicit,
but it is a pattern that is commonly used for singletons, e.g. loggers.
Use thoughtfully, and remember that injecting the export service
explicitly in some widgets is always possible!

Next steps: as soon as the export.Device is not longer necessary, then
the export.Service type doesn't need to be exported anymore.