Stop using furl
We needed to update furl for Python 3.10+ support, but after looking at
it, we can drop it entirely.

The valid_path() check to see if the path contains a hostname is
straightforward to switch to urllib.parse.urlparse().

In prep_request(), we use furl for validation and path normalization.
urlparse() is actually better for validation since it parses against the
URL spec[1]. But one weird part is that port validation only applies
when you actually access the `.port` member variable.

For the path normalization, I just copied the 3 lines (now 8) of code
out of furl. It should be very stable, it was never modified in furl
once it was added 9 years ago.

Finally, the test_proxy_500_misconfiguration test case had to be
adjusted, because it relied on furl incorrectly interpreting an extra
colon as an invalid port. The config now uses an explicitly invalid port
("-1") so it fails under the new code.

[1] https://url.spec.whatwg.org/

Fixes #105.