Fix flaky test!
Resetting the export.Service as part of the setup of the tests is
a compromise. It is a smell, that accurately points at other tests
not performing an adequate teardown.

I haven't fixed those this time around because the way they are written
doesn't provide an obvious place to teardown related tests, so adding
teardown operations introduces a significant amount of duplication.

It seems to me that athough less than ideal, this will prove more
reliable and more less error-prone for the time being.