Fix testing requirement checks: scope it to Debian version
The requirement files for Bullseye and Bookworm used ot only
differ in comments. That's not the case anymore, so this commit
scoped the check to the requirement file that matches
the Debian version under test.

I've verified that the check fails as expected when the file
is out of date.